[CHORE] #40 : manage -> management renaming
