+++ .idea/gradle.xml
@@ -36,7 +36,7 @@
             <option value="$PROJECT_DIR$/core/network" />
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/auth" />
-            <option value="$PROJECT_DIR$/feature/manage" />
+            <option value="$PROJECT_DIR$/feature/management" />
             <option value="$PROJECT_DIR$/feature/notice" />
             <option value="$PROJECT_DIR$/feature/profile" />
             <option value="$PROJECT_DIR$/feature/splash" />
