+++ app/src/main/res/drawable/ic_management.xml
