+++ app/src/main/res/drawable/ic_management_select.xml
