--- app/src/main/res/drawable/selector_manage.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:drawable="@drawable/ic_manage_select"
-        android:state_checked="true"/>
-    <item
-        android:drawable="@drawable/ic_manage"
-        android:state_checked="false"/>
-</selector>
