+++ app/src/main/res/drawable/selector_management.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:drawable="@drawable/ic_management_select"
+        android:state_checked="true"/>
+    <item
+        android:drawable="@drawable/ic_management"
+        android:state_checked="false"/>
+</selector>
