--- core/data/src/main/java/com/wap/wapp/core/data/repository/manage/ManageRepository.kt
@@ -1,9 +1,0 @@
-package com.wap.wapp.core.data.repository.manage
-
-interface ManageRepository {
-    suspend fun getManager(userId: String): Result<Boolean>
-
-    suspend fun postManager(userId: String): Result<Unit>
-
-    suspend fun getManageCode(code: String): Result<Boolean>
-}
