+++ core/data/src/main/java/com/wap/wapp/core/data/repository/management/ManagementRepository.kt
@@ -1,0 +1,9 @@
+package com.wap.wapp.core.data.repository.management
+
+interface ManagementRepository {
+    suspend fun getManager(userId: String): Result<Boolean>
+
+    suspend fun postManager(userId: String): Result<Unit>
+
+    suspend fun getManagementCode(code: String): Result<Boolean>
+}
