+++ core/network/src/main/java/com/wap/wapp/core/network/source/management/ManagementDataSource.kt
@@ -1,0 +1,9 @@
+package com.wap.wapp.core.network.source.management
+
+interface ManagementDataSource {
+    suspend fun getManager(userId: String): Result<Boolean>
+
+    suspend fun postManager(userId: String): Result<Unit>
+
+    suspend fun getManagementCode(code: String): Result<Boolean>
+}
