--- feature/manage/src/androidTest/java/com/wap/wapp/feature/manage/ExampleInstrumentedTest.kt
@@ -1,22 +1,0 @@
-package com.wap.wapp.feature.manage
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import junit.framework.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.wap.wapp.feature.manage.test", appContext.packageName)
-    }
-}
