+++ feature/management/src/main/AndroidManifest.xml
