+++ feature/management/src/main/res/navigation/nav_management.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_management"
+    app:startDestination="@id/manageFragment">
+
+    <fragment
+        android:id="@+id/manageFragment"
+        android:name="com.wap.wapp.feature.management.ManagementFragment"
+        android:label="fragment_manage" />
+
+    <deepLink app:uri="wapp://feature/nav_management"/>
+</navigation>
