+++ settings.gradle.kts
@@ -28,5 +28,5 @@ include(":feature:auth")
 include(":feature:notice")
 include(":feature:survey")
 include(":feature:profile")
-include(":feature:manage")
+include(":feature:management")
 include(":feature:splash")
